+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All feature additions, significant bug fixes, and API changes will be documented
 in this file. This project follows [semantic versioning](https://semver.org/).
 
+# 16.1
+
+- Add `Observable::queue` which creates and observable where upstream and downstream
+  run concurrently and are seperated by a queue of a given maximum size.
+
 # 16.0
 
 - Make `Fiber::run` and `Task::run` execute as much of the given task synchronously